Change canAdmin user checkbox to tri-state box.
When the user is not admin and in no admin team, checkbox is off.
When the user is admin, checkbox is on.
When the user is not admin but in an admin team, checkbox is indeterminate.

Add a file with test users to `src/test/resources` which displays
the different behaviour, also with external users.