Use explicit versions for JUnit and JaCoCo, JUnit update to 4.12
Use explicit coordinates, and therewith version number, for JUnit
and JaCoCo in the build.moxie file, instead of just `junit`.
It should not be some version that just happens to be used.

Update to JUnit version 4.12.

The same is true for JaCoCo, even more so. But updating to the latest
JaCoCo version brakes the build of the documentation somehow.
So the update must be done later, when there is a safe way to do so.
An update of JaCoCo is necessary to make it work under Java 8.